pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apitf"
version = "1.0.0"
description = "Multi-environment API test framework — REST Countries + Open-Meteo"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31,<3.0",
    "pyyaml>=6.0,<7.0",
    "pdfplumber>=0.10,<1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4,<9.0",
    "allure-pytest>=2.13,<3.0",
    "pytest-rerunfailures>=12.0,<16.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["apitf*"]

[tool.setuptools.package-dir]
"" = "."
